# 05_Application_Design/05_Application_Design/API_Specification.yaml
openapi: 3.0.3
info:
  title: Finished Goods Warehouse API Wrapper
  description: API wrapper gọi SQL Server Stored Procedure cho hệ thống WMS kho thành phẩm. Business logic nằm trong Stored Procedure.
  version: 0.6.0
servers:
  - url: https://wms-api.local/api/v1
    description: Internal API wrapper
  - url: http://localhost:5000/api/v1
    description: Local development
...
paths:
...
  /pack360/{pack360Id}/complete:
    post:
      tags: [Pack360]
      summary: Complete Pack360
      description: Wrapper gọi Stored Procedure usp_Pack360_Complete.
      operationId: completePack360
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: pack360Id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Kết quả xử lý từ Stored Procedure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
  /pack360/{pack360Id}/release:
    post:
      tags: [Pack360]
      summary: Giải phóng Pack360
      description: Wrapper gọi Stored Procedure usp_Pack360_Release.
      operationId: releasePack360
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: pack360Id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: Kết quả xử lý từ Stored Procedure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
...
components:
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: true
      schema:
        type: string
      description: Mã chống gửi lặp, được truyền xuống Stored Procedure.
  schemas:
    CommandResult:
      type: object
      properties:
        status:
          type: string
          enum: [SUCCESS, FAILED, DUPLICATED, CONFLICT, UNAUTHORIZED]
        message:
          type: string
        error_code:
          type: string
          nullable: true
        document_no:
          type: string
          nullable: true
        object_code:
          type: string
          nullable: true
        request_id:
          type: string
        trace_id:
          type: string
      required: [status, message, request_id]
...
    ReasonRequest:
      type: object
      properties:
        reason_code:
          type: string
        note:
          type: string
        user_code:
          type: string
      required: [reason_code, user_code]
...